Silence python warnings if verbosity is set to zero i.e. `make … verbosity=0`
